components: fix type linting
the duplicate line in `from_primitive` is an unfortunate side effect of mypy's limitations - it cannot infer that both
FullElement and DiffElement have a `from_primitive` method with 1 argument of the same type.
There is a way to fix this, but it makes the code thoroughly unreadable and impossible to maintain.